✨ include family and type in tokens and states
