+++ packages/atom.io/src/selector.ts
@@ -1,13 +1,9 @@
-import HAMT from "hamt_plus"
-import * as Rx from "rxjs"
-
-import { become } from "~/packages/anvl/src/function"
 import type { Serializable } from "~/packages/anvl/src/json"
 import { stringifyJson } from "~/packages/anvl/src/json"
 
-import type { ReadonlyValueToken, SelectorToken } from "."
-import type { Selector, Store } from "./internal"
-import { IMPLICIT, markDone, deposit, registerSelector } from "./internal"
+import type { FamilyMetadata, ReadonlyValueToken, SelectorToken } from "."
+import type { Store } from "./internal"
+import { selector__INTERNAL, IMPLICIT, deposit } from "./internal"
 import type { ReadonlyTransactors, Transactors } from "./transaction"
 
 export type SelectorOptions<T> = {
@@ -29,57 +25,7 @@ export function selector<T>(
   options: ReadonlySelectorOptions<T> | SelectorOptions<T>,
   store: Store = IMPLICIT.STORE
 ): ReadonlyValueToken<T> | SelectorToken<T> {
-  if (HAMT.has(options.key, store.selectors)) {
-    store.config.logger?.error(
-      `Key "${options.key}" already exists in the store.`
-    )
-  }
-
-  const subject = new Rx.Subject<{ newValue: T; oldValue: T }>()
-
-  const { get, set } = registerSelector(options.key, store)
-  const getSelf = () => {
-    const value = options.get({ get })
-    store.valueMap = HAMT.set(options.key, value, store.valueMap)
-    return value
-  }
-
-  if (!(`set` in options)) {
-    const readonlySelector = {
-      ...options,
-      subject,
-      get: getSelf,
-    }
-    store.readonlySelectors = HAMT.set(
-      options.key,
-      readonlySelector,
-      store.readonlySelectors
-    )
-    const initialValue = getSelf()
-    store.config.logger?.info(`   ✨ "${options.key}" =`, initialValue)
-    return { ...readonlySelector, type: `readonly_selector` }
-  }
-
-  const setSelf = (next: T | ((oldValue: T) => T)): void => {
-    store.config.logger?.info(`   <- "${options.key}" became`, next)
-    const oldValue = getSelf()
-    const newValue = become(next)(oldValue)
-    store.valueMap = HAMT.set(options.key, newValue, store.valueMap)
-    markDone(options.key, store)
-    subject.next({ newValue, oldValue })
-    options.set({ get, set }, newValue)
-  }
-
-  const mySelector: Selector<T> = {
-    ...options,
-    subject,
-    get: getSelf,
-    set: setSelf,
-  }
-  store.selectors = HAMT.set(options.key, mySelector, store.selectors)
-  const initialValue = getSelf()
-  store.config.logger?.info(`   ✨ "${options.key}" =`, initialValue)
-  return { ...mySelector, type: `selector` }
+  return selector__INTERNAL(options, undefined, store)
 }
 
 export type SelectorFamilyOptions<T, K extends Serializable> = {
@@ -105,7 +51,9 @@ export function selectorFamily<T, K extends Serializable>(
   store: Store = IMPLICIT.STORE
 ): (key: K) => ReadonlyValueToken<T> | SelectorToken<T> {
   return (key: K): ReadonlyValueToken<T> | SelectorToken<T> => {
-    const fullKey = `${options.key}__${stringifyJson(key)}`
+    const subKey = stringifyJson(key)
+    const family: FamilyMetadata = { key: options.key, subKey }
+    const fullKey = `${options.key}__${subKey}`
     const existing =
       store.selectors.get(fullKey) ?? store.readonlySelectors.get(fullKey)
     if (existing) {
@@ -116,18 +64,20 @@ export function selectorFamily<T, K extends Serializable>(
       get: options.get(key),
     }
     if (!(`set` in options)) {
-      return selector<T>(
+      return selector__INTERNAL<T>(
         {
           ...readonlySelectorOptions,
         },
+        family,
         store
       )
     }
-    return selector<T>(
+    return selector__INTERNAL<T>(
       {
         ...readonlySelectorOptions,
         set: options.set(key),
       },
+      family,
       store
     )
   }
